rip adapter
